thinkforge-demo-20250801/requirements.txt
sqlalchemy>=1.4,<2.0
sentence-transformers>=2.2.0
numpy>=1.18.0
typing-extensions>=4.0.0
pydantic>=1.9.0,<2.0.0
tqdm>=4.62.0
scikit-learn>=1.0.0
requests>=2.28.0
psycopg2-binary>=2.9.3
openai>=1.4.0
selenium>=4.11.2
# Add other direct dependencies if identified, e.g., specific DB drivers are NOT included here. 